requirements.txt
flask==2.2.3
werkzeug==2.2.3
Pillow==10.0.0
requests==2.31.0
gunicorn==21.2.0
torch>=1.10.0
torchvision>=0.11.0
transformers>=4.23.0
scikit-image>=0.19.0
numpy>=1.22.0 